test(tab-bar): migrate translucent
