+++ core/src/components/tab-bar/test/translucent/tab-bar.e2e.ts
@@ -1,0 +1,43 @@
+import { expect } from '@playwright/test';
+import { configs, test } from '@utils/test/playwright';
+
+/**
+ * Translucent is only available in iOS mode
+ */
+configs({ modes: ['ios'], directions: ['ltr'] }).forEach(({ title, screenshot, config }) => {
+  test.describe(title('tab-bar: translucent'), () => {
+    test.beforeEach(({ skip }) => {
+      skip.browser('firefox', 'Firefox does not support translucent effect');
+    });
+    test('should render translucent tab bar', async ({ page }) => {
+      await page.setContent(
+        `
+        <style>
+          body {
+            background: linear-gradient(to right, orange, yellow, green, cyan, blue, violet);
+          }
+        </style>
+        <ion-tab-bar translucent="true" selected-tab="1">
+          <ion-tab-button tab="1">
+            <ion-label>Recents</ion-label>
+          </ion-tab-button>
+
+          <ion-tab-button tab="2">
+            <ion-label>Favorites</ion-label>
+            <ion-badge>23</ion-badge>
+          </ion-tab-button>
+
+          <ion-tab-button tab="3">
+            <ion-label>Settings</ion-label>
+          </ion-tab-button>
+        </ion-tab-bar>
+      `,
+        configs
+      );
+
+      const tabBar = page.locator('ion-tab-bar');
+
+      await expect(tabBar).toHaveScreenshot(screenshot(`tab-bar-translucent`));
+    });
+  });
+});
